[refactor]: astro-reactive-form types
